[ANDROSDK-1154-DATA] Refactor to make metadata params variable
