+++ core/src/androidTest/java/org/hisp/dhis/android/localanalytics/LocalAnalyticsParams.kt
@@ -27,17 +27,17 @@
  */
 package org.hisp.dhis.android.localanalytics
 
-data class LocalAnalyticsParams(val organisationUnitChildren: Int,
-                                val categoryOptionCombos2: Int,
-                                val categoryOptionCombos3: Int,
-                                val dataElementsAggregated: Int,
-                                val dataElementsTracker: Int,
-                                val programStagesWithRegistration: Int,
-                                val programStagesWithoutRegistration: Int,
-                                val trackedEntityAttributes: Int) {
+data class LocalAnalyticsMetadataParams(val organisationUnitChildren: Int,
+                                        val categoryOptionCombos2: Int,
+                                        val categoryOptionCombos3: Int,
+                                        val dataElementsAggregated: Int,
+                                        val dataElementsTracker: Int,
+                                        val programStagesWithRegistration: Int,
+                                        val programStagesWithoutRegistration: Int,
+                                        val trackedEntityAttributes: Int) {
 
-    companion object LocalAnalyticsParams {
-        val Default = LocalAnalyticsParams(
+    companion object LocalAnalyticsMetadataParams {
+        val Default = LocalAnalyticsMetadataParams(
                 organisationUnitChildren = 3,
                 categoryOptionCombos2 = 2,
                 categoryOptionCombos3 = 6,
@@ -46,5 +46,13 @@ data class LocalAnalyticsParams(val organisationUnitChildren: Int,
                 programStagesWithRegistration = 3,
                 programStagesWithoutRegistration = 1,
                 trackedEntityAttributes = 10)
+    }
+}
+
+data class LocalAnalyticsDataParams(val trackedEntityInstances: Int)
+
+object LocalAnalyticsMetadataParamsGenerator {
+    fun generateMetadataParams(dataParams: LocalAnalyticsDataParams): LocalAnalyticsMetadataParams {
+        return LocalAnalyticsMetadataParams.Default
     }
 }
